Fix circular dependency between Eventing and commerce mock in the fast-integration tests
