+++ tests/fast-integration/eventing-test/utils.js
@@ -12,7 +12,7 @@ const {
 
 const {DirectorClient, DirectorConfig, getAlreadyAssignedScenarios} = require('../compass');
 const {GardenerClient, GardenerConfig} = require('../gardener');
-const fs = require('fs');
+const {eventMeshSecretFilePath} = require('./common/common');
 const isSKR = process.env.KYMA_TYPE === 'SKR';
 const skipResourceCleanup = process.env.SKIP_CLEANUP || false;
 const suffix = getSuffix(isSKR);
@@ -22,13 +22,9 @@ const testNamespace = `test-${suffix}`;
 const mockNamespace = process.env.MOCK_NAMESPACE || 'mocks';
 const backendK8sSecretName = process.env.BACKEND_SECRET_NAME || 'eventing-backend';
 const backendK8sSecretNamespace = process.env.BACKEND_SECRET_NAMESPACE || 'default';
-const eventMeshSecretFilePath = process.env.EVENTMESH_SECRET_FILE || '';
 const DEBUG_MODE = process.env.DEBUG;
 const timeoutTime = 10 * 60 * 1000;
 const slowTime = 5000;
-const natsBackend = 'nats';
-const bebBackend = 'beb';
-const eventMeshNamespace = getEventMeshNamespace();
 
 // SKR related constants
 let gardener = null;
@@ -38,16 +34,6 @@ if (isSKR) {
   gardener = new GardenerClient(GardenerConfig.fromEnv()); // create gardener client
   director = new DirectorClient(DirectorConfig.fromEnv()); // director client for Compass
   shootName = getShootNameFromK8sServerUrl();
-}
-
-// reads the EventMesh namespace from the credentials file
-function getEventMeshNamespace() {
-  try {
-    const eventMeshSecret = JSON.parse(fs.readFileSync(eventMeshSecretFilePath, {encoding: 'utf8'}));
-    return '/' + eventMeshSecret['namespace'];
-  } catch (e) {
-    return undefined;
-  }
 }
 
 // cleans up all the test resources including the compass scenario
@@ -105,7 +91,6 @@ module.exports = {
   isSKR,
   backendK8sSecretName,
   backendK8sSecretNamespace,
-  eventMeshSecretFilePath,
   DEBUG_MODE,
   timeoutTime,
   slowTime,
@@ -114,8 +99,5 @@ module.exports = {
   shootName,
   suffix,
   cleanupTestingResources,
-  natsBackend,
-  bebBackend,
-  eventMeshNamespace,
   getRegisteredCompassScenarios,
 };
